requirements.txt
python-telegram-bot==20.7
websockets==12.0
aiohttp==3.9.3
python-dotenv==1.0.1
asyncpg==0.31.0
